Extracting ClientForAdditionalAPIs
