--- repo-client/src/main/java/io/lionweb/repoclient/TransferFormat.java
@@ -1,7 +1,0 @@
-package io.lionweb.repoclient;
-
-public enum TransferFormat {
-  JSON,
-  PROTOBUF,
-  FLATBUFFERS
-}
